add projects managment
Allow user to create and modify FragMAX project.

Each project is the specification of the protein, fragments
library, proposal number and where experiment data is located.

The experiment data is specified by a list of shifts when the
experiment data was collected.

Adds a drop-down list which allows the use to select 'current project'.

The 'current project' decides which data is display and processed by
our application.
prototype for project switching GUI

prototype for projects list page

manage project machinery