include update_status.py script into the repository
The update_status.py script is run after each processing
job on HPC is done. It will update the allstatus.csv file,
to reflect that new result are available for dataset.

When a new project is created, copy the update_status.py
script to project's fragmax/script directory, and use
that version of the script from HPC.

Use python3 to run the script, using some additional 'module load...'
magic. We need to use python3, as the script is using the
interpolated string expressons, e.g. f"{foo} {bar}"