create project status CSV file when project is created
Create the initial version of the 'allstatus.csv' file
when project is created. Otherwise it created when user
clicked on 'home' for the first time of a new project.

It can take a while for the 'allstatus.csv' to be created,
thus the site can become unresponsive on first 'home' click.