use DIALS 2.1.3 and add memory size requests
To solve issues with where prossessing of some datasets failed
due to OOM-killer being activated, use DIALS version 2.1.3 and
request minimum memory size for the task.

The older DIALS had some memory usage bug, which is solved in 2.1.3.

Empirically, we know that we need somewere around 210G of memory
to process some datasets, when we do a 48-way parallelization.

When no '--mem' flag is specifed, some jobs may fail, even if
DIALS 2.1.3 is used.