refactor the 'Home' view
a new module that manages reading of the 'allstatus.csv'
and 'datacollections.csv'

the dset.get_datasets() function returns a list of
DataSet objects. Each DataSet object contains the
dataset's metadata and processing status.

Move the generation of the HTML of each dataset's processing
status into the datasets.html template.