on 'Home' page, lazy-load crystal and fragment images
Load the crystal snapshot images and fragment svg graphics
only when the respective table cell becomes visible.

This make it faster to initially load the 'Home' page make
it possible to work with project with large ammount of
dataset.

Also, cuts down on number of request and network traffic in
many cases, as most of the dataset will not be views each
time the page is loaded.