use jQuery style to handle custom params for dataproc
Use the jQuery calls to pass on custom params for data
processing tools.

This way we transparantly handle tools that are not present
at some deployment site.

For example, if a site does not provide autoPROC, the
autoPROC 'custom parameter' input element will not be
shown. The jQuery will gives us 'undefined' value for
that value, which will be posted as empty string in
the POST request.