density view: only load ligfit/rhofit PDB when available
Use the template variables 'show_ligfit' and 'show_rhofit' as conditionals
to load ligfit/rhofit generated PDBs.

This way, we can specify what kind of result the density view is presenting,
and avoid trying to load PDBs for results that lack the respective PDB.

Also use 'show_ligfit' and 'show_rhofit' to implement the logic for when
we are rendering LigFit, RhoFit and 'open dual viewer' controls.