jobsd: handle large 'start_jobs' commands
Fixes issue where were not able to parse a 'start_jobs' command
with large 'jobs_set'.

Use custom code to read a line from the jobsd socket, as the
standard StreamReader.readline() have a limit on how long lines
it can handle.

As we serialize the whole command json to a single line, we can
easily hit that limit, when for example starting a job to process
200-300 sets in one command.