jobsd: 'dynamic' poll timeouts
Poll with different intervalls, depending if
there is any recent activity.

First poll with short perioids, then with
longer. If any activity is detected, such
as new commands issued, job started or stopped,
start polling with short periods.

This mainly to avoid hammering the SLURM front-end
host to much with 'sacct' command if we have
long-running jobs, which no user is activly
watching.