redis lock: release lock on exception
If the redis lock using code have raised an
exception, make sure we release the lock.

Without the try-finally block, in cases when
the exception is raised, the lock was not relased.

This means in pratical terms, if some of the tasks,
which use locking, were crashing, you'll need
to restart redis to get rid if the lock.