disable 'autoPROC' on data analysis page
Disable the feature of re-processing data with autoPROC
for now.

Now, that we don't copy autoPROC files inside the 'fragmax'
folder, we need to keep track if the user have reprocessed
the dataset with autoPROC.

So that when we run the refine tools, we need to be able to
pick-up either the original MTZ file, or the new re-processed
by the user.

Until that stuff is in place, disallow to run autoPROC from
the within the app.