jobsd: throttle jobs by counting CPUs
For data processing, refine and ligfit jobs, keep track of
how many CPU the job is expected to use.

On HZB, start jobds with --cpu-limit flag. The flag
enables CPU based throttling.

That is, we count the CPU required for each launched jobs,
and when we reach the specified limit, we wait until previous
jobs are finished and required ammount of CPUs is available.