xia2 scraper: handle absent log directory
When generating a list of logs file, handle the case where
the job fails before creating 'LogFiles' directory.

This can for example happen when xia2/dials is invoked with arguments
that it does not understand.

We'll get an 'process result' entry for that job in project's DB. But
the job will exit before creating 'LogFiles' directory.