PDB: store space group and cell vals in DB
When a PDB file is added to the project, store PDB file's
space group and unit cell parameters in the project database.

We want to store these values in the DB for performance reasons
mainly. Parsing PDBs is a bit slow operation. Having these
values in the DB allows us to quickly access them for example
when checking if PDB is 'compatible' with some MTZ file.