remove job's old files
Make sure that in cases when we re-run a job with the same
tool + input data, previously generated files are removed.

This solves the issue with for example xdsapp is being re-run
with space group and cell params specified, the new space group
was not reflectd in the projects database.