store dataset's beamline in database
When setting up project, use dataset's meta-data to
figure out the beamline where it was collected.

Store the beamline for each dataset in the database.

Besides showing this in the UI, it is also potentionally
usefull when making depositions to PDB. PDB requires that
the beamline where the original data was collected to be
specified.