exclude failed refine results from PanDDa
When picking the refined datasets to include into a PanDDa analysis
run, filter out failed runs.

This extra filtering need to be performed to handle the following
situation.

Imagine dataset 'Foo' was once successfully refined. An 'RefineResult'
entry will be added to the database. At some later point dataset
'Foo' is refined again, but this time the refinement failed. The
'RefineResult' entry will still exist in the database, but now with
'error' status.

The file generated by the last failed run are potentionally mangled or
missing, thus we need to exclude them from PanDDa run.